requirements.txt
streamlit
faiss-cpu
sentence-transformers
pypdf
python-dotenv
openai
requests
beautifulsoup4
numpy